Fix offsets for return annotation checks
Calculate line number as the line before where the body of the function begins rather than where the function definition starts which AST provides as the line where the decorators start.

Use the source code to determine where the function definition ends by searching for the index of ":" in the line where we believe the definition ends. If ":" cannot be found for whatever reason, column offset will default to the function node's column offset.

Closes: #21
Closes: #23