Add parsing support for the `--suppress-none-returning` config flag
With the `--suppress-none-returning` flag set, TYP200 level errors will be suppressed for functions that have no `return` statements, or explicit `return` statements all return `None`.

A special-case `ast.NodeVisitor` class is provided to visit the `Return` nodes of the function passed to the `Function.from_function_node` classmethod, and setting a flag if the function returns anything other than None.