Update Azure config to use flake8 entry point
There appears to be an upstream issue with the entrypoints library (via flake8) that causes the option parser to double register custom flake8 configuration options when invoking as a module vs. the flake8 entrypoint. While the issue is fixed on flake8 master, until a new version is released we can utilize the flake8 entry point to resolve the CI issue.

See: gitlab.com/pycqa/flake8/issues/610