Fix incorrect return column index on lines with existing annotations
Using `str.find()` to locate the colon denoting the end of the function definition fails in the presence of other annotations in the same line of source code, as it identifies the location of the first colon which isn't necessarily the close of the function definition. As this closing colon should be the last instance for a line of source code, changing to `str.rfind()` should mitigate this issue.