Implement context switcher to fix misclassification of nested returns
The previous ReturnVisitor being used to identify whether or not a function returns anything other than None was flawed in its implementation, in that it visited all return nodes, even those in nested functions, and incorrectly used them in its determination about whether or not the top-level node returns something other than None.

To fix this issue, a context switcher is implemented to keep track of the function node that contains the non-None return node. If node(s) are found, they are added to a set. The `has+only_none_returns` attribute is replaced by a property method that checks for the top-level node's presence in this set of nodes.