Add flag for suppressing all errors from dynamically typed functions
To help facilitate the gradual addition of typing to existing codebases, Mypy by default will not type check dynamically typed functions (aka functions with no hints). Because our default behavior is solely to check for the presence of annotations, this can create a conflict when the lack of hints for a particular function is intentional.

To support the use case of gradual addition of typing to a codebase, this introduces an `--allow-untyped-defs` flag to optionally suppress all errors from completely unannotated function definitions, similar to how Mypy provides the `--disallow-untyped-defs` flag to suppress the default behavior of ignoring dynamically typed functions.