Fix bug in colon seeker for multiline docstrings in Py < 3.8
Multiline docstrings in Python versions prior to 3.8 require rewinding through the body to seek the opening of the multiline docstring, which is the line after the close of the function def. The previous logic contained an indexing bug where it rewound prematurely at the beginning. This wasn't a problem for multiline docstrings with more than one line of content, but if the multiline docstring is empty or contained only one file then this jumped over the opening of the docstring and just keeps rewinding until an unrelated triple quote is found (or even all the way around to the starting point).

This indexing has been adjusted & the comments changed to better explain the flow of the seeker.

Off-by-one errors, you'll never be safe.